+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.1.1
+=====
+
+* (bug) Remove debug output div
+
+
 2.1.0
 =====